requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
paramiko==3.4.0
qrcode[pil]==7.4.2
Flask==3.0.0
requests==2.31.0
yoomoney==0.1.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.20
tenacity==8.2.3

